Reverted bh_front, and moved digital display code to its own class
