+++ aphavoc/source/gunships/avionics/blackhawk/bh_avion.h
@@ -74,7 +74,7 @@
 
 #include "bh_front.h"
 
-//#include "bh_chrono.h"
+#include "bh_digits.h"
 
 #include "bh_hms.h"
