+++ aphavoc/source/gunships/avionics/blackhawk/bh_front.c
@@ -73,9 +73,7 @@
 #define VIEWPORT_WIDTH		(84.0)
 #define VIEWPORT_HEIGHT		(40.0)
 
-#define LINE_LENGTH  36
-
-#define MAX_STRING_LENGTH	(12)
+#define MAX_STRING_LENGTH	(14)
 
 /////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 /////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
@@ -86,385 +84,66 @@ static rgb_colour
 	clear_colour;
 
 static screen
-	*ng_screen,
-	*tgt_screen,
-	*trq_screen,
-	*fuel_quantity_screen,
-	*chaff_flare_screen,
-	*rdr_alt_screen,
-	*chrono_screen;
+	*texture_screen;
 
 static char
 	line1[MAX_STRING_LENGTH + 1],
-	line2[MAX_STRING_LENGTH + 1];
+	line2[MAX_STRING_LENGTH + 1],
+	line3[MAX_STRING_LENGTH + 1],
+	line4[MAX_STRING_LENGTH + 1];
 
 /////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 /////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 /////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 
-//static void display_messages (float x_org, float y_org)
-//{
-//	float y = 1.0;
-//	int i=0;
-//	char buffer[64];
-//
-//	set_mono_font_type(MONO_FONT_TYPE_6X10);
-//	set_mono_font_colour(text_colour);
-//
-//	for (i=0; i<5; i++)
-//	{
-//		snprintf(buffer, LINE_LENGTH+1, "%-11s|%-11s|%-12s",
-//			warnings[(last_warning + i) % 5], cautions[(last_caution + i) % 5], advisories[(last_advisory + i) % 5]);
-//
-//		set_mono_font_position(0.0, y);
-//		set_mono_font_rel_position(0.0, 0.0);
-//
-//		print_mono_font_string(buffer);
-//		y += 10;
-//	}
-//
-//	for (i=0; i<4; i++)
-//	{
-//		set_mono_font_position(0.0, y);
-//		set_mono_font_rel_position(0.0, 0.0);
-//
-//		print_mono_font_string(radio_freqs[i]);
-//		y += 10;
-//	}
-//
-//	{
-//		float hours, minutes, seconds;
-//		float time_of_day = get_local_entity_float_value (get_session_entity (), FLOAT_TYPE_TIME_OF_DAY);
-//		get_digital_clock_values (time_of_day, &hours, &minutes, &seconds);
-//
-//		snprintf(buffer, LINE_LENGTH+1, "%s %02d:%02d:%02d L", radio_freqs[4], (int) hours, (int) minutes, (int) seconds);
-//
-//		set_mono_font_position(0.0, y);
-//		set_mono_font_rel_position(0.0, 0.0);
-//
-//		print_mono_font_string(buffer);
-//	}
-//#if 0
-//	float
-//		x_min,
-//		y_min,
-//		x_max,
-//		y_max,
-//		y_line1,
-//		y_line2,
-//		y_line3,
-//		y_line4;
-//
-//	if (!electrical_system_active())
-//		return;
-//
-//	x_min = x_org;
-//	y_min = y_org;
-//	x_max = x_org + VIEWPORT_WIDTH;
-//	y_max = y_org + VIEWPORT_HEIGHT;
-//
-//	set_viewport (x_min, y_min, x_max, y_max);
-//
-//	y_line1 = y_org;
-//	y_line2 = y_org + 10.0;
-//	y_line3 = y_org + 20.0;
-//	y_line4 = y_org + 30.0;
-//
-//	set_mono_font_type (MONO_FONT_TYPE_5X9);
-//
-//	set_mono_font_colour (text_colour);
-//
-//	set_mono_font_position (x_min, y_line1);
-//
-//	print_mono_font_string (line1);
-//
-//	set_mono_font_position (x_min, y_line2);
-//
-//	print_mono_font_string (line2);
-//
-//	set_mono_font_position (x_min, y_line3);
-//
-//	print_mono_font_string (line3);
-//
-//	set_mono_font_position (x_min, y_line4);
-//
-//	print_mono_font_string (line4);
-//#endif
-//}
-
-/////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-/////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-/////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-
-static void display_ng (void)
+static void display_messages (float x_org, float y_org)
 {
-	float y = 1.0;
-	char buffer[64];
+#ifndef OGRE_EE
+	float
+		x_min,
+		y_min,
+		x_max,
+		y_max,
+		y_line1,
+		y_line2,
+		y_line3,
+		y_line4;
 
-	set_mono_font_type(MONO_FONT_TYPE_17X26_DIGITAL);
-	set_mono_font_colour(text_colour);
+	if (!electrical_system_active())
+		return;
 
-	sprintf (buffer, "%d %d", (int) (bound(current_flight_dynamics->left_engine_n1_rpm.value, 0.0, 120.0)*10),
-			(int) (bound(current_flight_dynamics->right_engine_n1_rpm.value, 0.0, 120.0)*10));
+	x_min = x_org;
+	y_min = y_org;
+	x_max = x_org + VIEWPORT_WIDTH - 0.001;
+	y_max = y_org + VIEWPORT_HEIGHT - 0.001;
 
-	set_mono_font_position(0.0, y);
-	set_mono_font_rel_position(0.0, 0.0);
+	set_viewport (x_min, y_min, x_max, y_max);
 
-	print_mono_font_string(buffer);
-}
+	y_line1 = y_org;
+	y_line2 = y_org + 10.0;
+	y_line3 = y_org + 20.0;
+	y_line4 = y_org + 30.0;
 
-/////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-/////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-/////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
+	set_mono_font_type (MONO_FONT_TYPE_5X9);
 
-static void display_tgt (void)
-{
-	float y = 1.0;
-	char buffer[64];
+	set_mono_font_colour (text_colour);
 
-	set_mono_font_type(MONO_FONT_TYPE_17X26_DIGITAL);
-	set_mono_font_colour(text_colour);
+	set_mono_font_position (x_min, y_line1);
 
-	sprintf (buffer, "%d %d", (int) (bound(current_flight_dynamics->left_engine_temp.value, 0.0, 1000.0)),
-			(int) (bound(current_flight_dynamics->right_engine_temp.value, 0.0, 1000.0)));
-
-	set_mono_font_position(0.0, y);
-	set_mono_font_rel_position(0.0, 0.0);
-
-	print_mono_font_string(buffer);
-}
-
-/////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-/////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-/////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-
-static void display_trq (void)
-{
-	float y = 1.0;
-	char buffer[64];
-
-	set_mono_font_type(MONO_FONT_TYPE_17X26_DIGITAL);
-	set_mono_font_colour(text_colour);
-
-	sprintf (buffer, "%02d %02d", (int) (bound(current_flight_dynamics->left_engine_torque.value, 0.0, 120.0)),
-							(int) (bound(current_flight_dynamics->right_engine_torque.value, 0.0, 120.0)));
-
-	set_mono_font_position(0.0, y);
-	set_mono_font_rel_position(0.0, 0.0);
-
-	print_mono_font_string(buffer);
-}
-
-/////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-/////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-/////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-
-static void display_fuel (void)
-{
-	float y = 1.0;
-	char buffer[64];
-
-	set_mono_font_type(MONO_FONT_TYPE_17X26_DIGITAL);
-	set_mono_font_colour(text_colour);
-
-	sprintf (buffer, "%d", (int) (bound (kilograms_to_pounds (current_flight_dynamics->fuel_weight.value), 0.0, 2500.0)));
-
-	set_mono_font_position(0.0, y);
-	set_mono_font_rel_position(-1.0, 0.0);
-
-	print_mono_font_string(buffer);
-}
-
-/////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-/////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-/////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-
-static void display_chaff_flare (void)
-{
-	float y = 1.0;
-	char buffer[64];
-
-	int
-		lh_chaff,
-		rh_chaff,
-		lh_chaff_number,
-		rh_chaff_number,
-		total_chaff,
-		lh_flare,
-		rh_flare,
-		lh_flare_number,
-		rh_flare_number,
-		total_flare,
-		damaged;
-
-	entity
-		*en;
-
-	entity_sub_types
-		weapon_sub_type;
-
-	en = get_gunship_entity ();
-
-	set_mono_font_type(MONO_FONT_TYPE_17X26_DIGITAL);
-	set_mono_font_colour(text_colour);
-
-	lh_chaff = get_local_entity_weapon_hardpoint_info (en, BLACKHAWK_LHS_CHAFF_DISPENSER, ENTITY_SUB_TYPE_WEAPON_CHAFF, &weapon_sub_type, &lh_chaff_number, &damaged);
-	rh_chaff = get_local_entity_weapon_hardpoint_info (en, BLACKHAWK_RHS_CHAFF_DISPENSER, ENTITY_SUB_TYPE_WEAPON_CHAFF, &weapon_sub_type, &rh_chaff_number, &damaged);
-	lh_flare = get_local_entity_weapon_hardpoint_info (en, BLACKHAWK_LHS_FLARE_DISPENSER, ENTITY_SUB_TYPE_WEAPON_FLARE, &weapon_sub_type, &lh_flare_number, &damaged);
-	rh_flare = get_local_entity_weapon_hardpoint_info (en, BLACKHAWK_RHS_FLARE_DISPENSER, ENTITY_SUB_TYPE_WEAPON_FLARE, &weapon_sub_type, &rh_flare_number, &damaged);
-
-	total_chaff = lh_chaff_number + rh_chaff_number;
-	total_flare = lh_flare_number + rh_flare_number;
-
-	if (lh_chaff || rh_chaff || lh_flare || rh_flare)
-	{
-		if (!damaged)
-		{
-			sprintf (buffer, "%d          %d", total_chaff, total_flare);
-		}
-		else if (damaged)
-			{
-				if (blackhawk_damage.lh_chaff_dispenser)
-				{
-					total_chaff = rh_chaff_number;
-				}
-
-				if (blackhawk_damage.rh_chaff_dispenser)
-				{
-					total_chaff = lh_chaff_number;
-				}
-
-				if (blackhawk_damage.lh_flare_dispenser)
-				{
-					total_flare = rh_flare_number;
-				}
-
-				if (blackhawk_damage.rh_flare_dispenser)
-				{
-					total_flare = lh_flare_number;
-				}
-
-				if (blackhawk_damage.lh_chaff_dispenser && blackhawk_damage.rh_chaff_dispenser
-						&& blackhawk_damage.lh_flare_dispenser && blackhawk_damage.rh_flare_dispenser)
-				{
-					sprintf (buffer, "XX          XX");
-				}
-				else if (blackhawk_damage.lh_chaff_dispenser && blackhawk_damage.rh_chaff_dispenser)
-				{
-					sprintf (buffer, "XX          %d", total_flare);
-				}
-				else if (blackhawk_damage.lh_flare_dispenser && blackhawk_damage.rh_flare_dispenser)
-				{
-					sprintf (buffer, "%d          XX", total_chaff);
-				}
-				else
-				{
-					sprintf (buffer, "%d          %d", total_chaff, total_flare);
-				}
-			}
-	}
-
-	set_mono_font_position(0.0, y);
-	set_mono_font_rel_position(15.0, 0.0);
-
-	print_mono_font_string(buffer);
-}
-
-/////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-/////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-/////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-
-static void display_rdr_alt (void)
-{
-	float y = 1.0;
-	char buffer[64];
-
-	set_mono_font_type(MONO_FONT_TYPE_17X26_DIGITAL);
-	set_mono_font_colour(text_colour);
-
-	sprintf (buffer, "%d", (int) (bound(feet (current_flight_dynamics->radar_altitude.value), 0.0, 1500.0)));
-
-	set_mono_font_position(0.0, y);
-	set_mono_font_rel_position(0.0, 0.0);
-
-	print_mono_font_string(buffer);
-}
-
-/////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-/////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-/////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-
-static void display_chrono (void)
-{
-	float y = 10.0;
-
-	set_mono_font_type(MONO_FONT_TYPE_17X26_DIGITAL);
-	set_mono_font_colour(text_colour);
-
-	set_mono_font_position(0.0, y);
-
-	set_mono_font_rel_position(5.0, 0.0);
 	print_mono_font_string (line1);
 
-	set_mono_font_rel_position (12.0, 60.0);
+	set_mono_font_position (x_min, y_line2);
+
 	print_mono_font_string (line2);
-}
 
-/////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-/////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-/////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
+	set_mono_font_position (x_min, y_line3);
 
-void update_chrono (void)
-{
-	float
-		time,
-		mission_time;
+	print_mono_font_string (line3);
 
-	int
-		hours,
-		minutes,
-		seconds,
-		mission_hours,
-		mission_minutes,
-		mission_seconds;
+	set_mono_font_position (x_min, y_line4);
 
-	char
-		chrono_line1[64],
-		chrono_line2[64];
-
-	entity
-		*task;
-
-	//
-	// Mission elapsed time
-	//
-
-	task = get_player_task (NULL);
-	if (task)
-	{
-		ASSERT (get_local_entity_int_value (task, INT_TYPE_VISIBLE_TASK));
-
-		mission_time = get_local_entity_float_value (task, FLOAT_TYPE_ELAPSED_MISSION_TIME);
-
-		get_digital_clock_int_values (mission_time, &mission_hours, &mission_minutes, &mission_seconds);
-
-		sprintf (chrono_line1, "%02d:%02d:%02d", mission_hours, mission_minutes, mission_seconds);
-	} else
-	{
-		sprintf (chrono_line1, "--:--:--");
-	}
-
-	//
-	// Local time
-	//
-
-	time = get_local_entity_float_value (get_session_entity (), FLOAT_TYPE_TIME_OF_DAY);
-
-	get_digital_clock_int_values (time, &hours, &minutes, &seconds);
-
-	sprintf (chrono_line2, "%02d:%02d", hours, minutes);
-
-	set_blackhawk_upfront_display_text (chrono_line1, chrono_line2);
+	print_mono_font_string (line4);
+#endif
 }
 
 /////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
@@ -473,23 +152,16 @@ void update_chrono (void)
 
 void initialise_blackhawk_upfront_display (void)
 {
-	ng_screen = create_system_texture_screen (125, 26, TEXTURE_INDEX_HOKUM_COCKPIT_MFD_RHS_1, TEXTURE_TYPE_SINGLEALPHA);
+	texture_screen = create_system_texture_screen (128, 128, TEXTURE_INDEX_AVCKPT_DISPLAY_UPFRONT, TEXTURE_TYPE_SINGLEALPHA);
 
-	tgt_screen = create_system_texture_screen (125, 26, TEXTURE_INDEX_HOKUM_COCKPIT_MFD_LHS_1, TEXTURE_TYPE_SINGLEALPHA);
+	line1[0] = '\0';
+	line2[0] = '\0';
+	line3[0] = '\0';
+	line4[0] = '\0';
 
-	trq_screen = create_system_texture_screen (125, 26, TEXTURE_INDEX_AVCKPT_DISPLAY_LHS_MFD, TEXTURE_TYPE_SINGLEALPHA);
+	set_rgb_colour (text_colour, 0, 255, 0, 255);
 
-	fuel_quantity_screen = create_system_texture_screen (69, 26, TEXTURE_INDEX_AVCKPT_DISPLAY_UPFRONT, TEXTURE_TYPE_SINGLEALPHA);
-
-	chaff_flare_screen = create_system_texture_screen (140, 26, TEXTURE_INDEX_COMANCHE_MFD1, TEXTURE_TYPE_SINGLEALPHA);
-
-	rdr_alt_screen = create_system_texture_screen (85, 26, TEXTURE_INDEX_COMANCHE_MFD4, TEXTURE_TYPE_SINGLEALPHA);
-
-	chrono_screen = create_system_texture_screen (157, 104, TEXTURE_INDEX_HOKUM_COCKPIT_MFD_LHS_2, TEXTURE_TYPE_SINGLEALPHA);
-//TEXTURE_INDEX_HOKUM_COCKPIT_MFD_LHS_2
-	set_rgb_colour (text_colour, 65, 211, 42, 255);
-
-	set_rgb_colour (clear_colour, 180, 240, 0, 0);
+	set_rgb_colour (clear_colour, 0, 255, 0, 0);
 }
 
 /////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
@@ -498,13 +170,7 @@ void initialise_blackhawk_upfront_display (void)
 
 void deinitialise_blackhawk_upfront_display (void)
 {
-	destroy_screen (ng_screen);
-	destroy_screen (tgt_screen);
-	destroy_screen (trq_screen);
-	destroy_screen (fuel_quantity_screen);
-	destroy_screen (chaff_flare_screen);
-	destroy_screen (rdr_alt_screen);
-	destroy_screen (chrono_screen);
+	destroy_screen (texture_screen);
 }
 
 /////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
@@ -513,22 +179,40 @@ void deinitialise_blackhawk_upfront_display (void)
 
 void update_blackhawk_upfront_display (void)
 {
-	update_chrono ();
+	char
+		s1[80],
+		s2[80];
+
+	float
+		time_of_day,
+		hours,
+		minutes,
+		seconds;
+
+	time_of_day = get_local_entity_float_value (get_session_entity (), FLOAT_TYPE_TIME_OF_DAY);
+
+	get_digital_clock_values (time_of_day, &hours, &minutes, &seconds);
+
+	sprintf (s1, "%02d:%02d:%02d", (int) hours, (int) minutes, (int) seconds);
+
+	sprintf (s2, "FUEL %04d LBS", (int) (bound (kilograms_to_pounds (current_flight_dynamics->fuel_weight.value), 0.0, 2500.0)));
+
+	set_blackhawk_upfront_display_text (NULL, NULL, s1, s2);
 }
 
 /////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 /////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 /////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 
-//void draw_blackhawk_upfront_display_on_cockpit (float x_org, float y_org)
-//{
-//	if (lock_screen (active_screen))
-//	{
-//		display_messages (x_org, y_org);
-//
-//		unlock_screen (active_screen);
-//	}
-//}
+void draw_blackhawk_upfront_display_on_cockpit (float x_org, float y_org)
+{
+	if (lock_screen (active_screen))
+	{
+		display_messages (x_org, y_org);
+
+		unlock_screen (active_screen);
+	}
+}
 
 /////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 /////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
@@ -536,170 +220,15 @@ void update_blackhawk_upfront_display (void)
 
 void draw_blackhawk_upfront_display_on_texture (void)
 {
-	draw_ng_display_on_texture();
-	draw_tgt_display_on_texture();
-	draw_trq_display_on_texture();
-	draw_fuel_display_on_texture();
-	draw_chaff_flare_display_on_texture();
-	draw_rdr_alt_display_on_texture ();
-	draw_chrono_display_on_texture ();
-}
+	set_active_screen (texture_screen);
 
-/////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-/////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-/////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-
-void draw_ng_display_on_texture (void)
-{
-	set_active_screen (ng_screen);
-
-	if (lock_screen (ng_screen))
+	if (lock_screen (texture_screen))
 	{
-		set_block (0, 0, (ng_screen->width-1), (ng_screen->height-1), clear_colour);
+		set_block (0, 0, 127, 127, clear_colour);
 
-		display_ng ();
+		display_messages (0.0, 0.0);
 
-		unlock_screen (ng_screen);
-	}
-
-	set_active_screen (video_screen);
-}
-
-/////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-/////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-/////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-
-//void draw_right_ng_display_on_texture (void)
-//{
-//	set_active_screen (right_ng_screen);
-//
-//	if (lock_screen (right_ng_screen))
-//	{
-//		set_block (0, 0, (right_ng_screen->width-1), (right_ng_screen->height-1), clear_colour);
-//
-//		display_right_ng ();
-//
-//		unlock_screen (right_ng_screen);
-//	}
-//
-//	set_active_screen (video_screen);
-//}
-
-/////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-/////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-/////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-
-void draw_tgt_display_on_texture (void)
-{
-	set_active_screen (tgt_screen);
-
-	if (lock_screen (tgt_screen))
-	{
-		set_block (0, 0, (tgt_screen->width-1), (tgt_screen->height-1), clear_colour);
-
-		display_tgt ();
-
-		unlock_screen (tgt_screen);
-	}
-
-	set_active_screen (video_screen);
-}
-
-/////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-/////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-/////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-
-void draw_trq_display_on_texture (void)
-{
-	set_active_screen (trq_screen);
-
-	if (lock_screen (trq_screen))
-	{
-		set_block (0, 0, (trq_screen->width-1), (trq_screen->height-1), clear_colour);
-
-		display_trq ();
-
-		unlock_screen (trq_screen);
-	}
-
-	set_active_screen (video_screen);
-}
-
-/////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-/////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-/////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-
-void draw_fuel_display_on_texture (void)
-{
-	set_active_screen (fuel_quantity_screen);
-
-	if (lock_screen (fuel_quantity_screen))
-	{
-		set_block (0, 0, (fuel_quantity_screen->width-1), (fuel_quantity_screen->height-1), clear_colour);
-
-		display_fuel ();
-
-		unlock_screen (fuel_quantity_screen);
-	}
-
-	set_active_screen (video_screen);
-}
-
-/////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-/////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-/////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-
-void draw_rdr_alt_display_on_texture (void)
-{
-	set_active_screen (rdr_alt_screen);
-
-	if (lock_screen (rdr_alt_screen))
-	{
-		set_block (0, 0, (rdr_alt_screen->width-1), (rdr_alt_screen->height-1), clear_colour);
-
-		display_rdr_alt ();
-
-		unlock_screen (rdr_alt_screen);
-	}
-
-	set_active_screen (video_screen);
-}
-
-/////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-/////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-/////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-
-void draw_chaff_flare_display_on_texture (void)
-{
-	set_active_screen (chaff_flare_screen);
-
-	if (lock_screen (chaff_flare_screen))
-	{
-		set_block (0, 0, (chaff_flare_screen->width-1), (chaff_flare_screen->height-1), clear_colour);
-
-		display_chaff_flare ();
-
-		unlock_screen (chaff_flare_screen);
-	}
-
-	set_active_screen (video_screen);
-}
-
-/////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-/////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-/////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-
-void draw_chrono_display_on_texture (void)
-{
-	set_active_screen (chrono_screen);
-
-	if (lock_screen (chrono_screen))
-	{
-		set_block (0, 0, (chrono_screen->width-1), (chrono_screen->height-1), clear_colour);
-
-		display_chrono ();
-
-		unlock_screen (chrono_screen);
+		unlock_screen (texture_screen);
 	}
 
 	set_active_screen (video_screen);
@@ -713,7 +242,7 @@ void draw_chrono_display_on_texture (void)
 // note: NULL == do not overwrite, "" == overwrite with blank message
 //
 
-void set_blackhawk_upfront_display_text (char *s1, char *s2)
+void set_blackhawk_upfront_display_text (char *s1, char *s2, char *s3, char *s4)
 {
 	if (s1)
 	{
@@ -728,13 +257,22 @@ void set_blackhawk_upfront_display_text (char *s1, char *s2)
 
 		line2[MAX_STRING_LENGTH] = '\0';
 	}
+
+	if (s3)
+	{
+		strncpy (line3, s3, MAX_STRING_LENGTH);
+
+		line3[MAX_STRING_LENGTH] = '\0';
+	}
+
+	if (s4)
+	{
+		strncpy (line4, s4, MAX_STRING_LENGTH);
+
+		line4[MAX_STRING_LENGTH] = '\0';
+	}
 }
 
 /////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 /////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 /////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-
-void clear_blackhawk_upfront_display(void)
-{
-
-}
